requirements.txt
requests==2.31.0
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyyaml==6.0.1
pyarrow==14.0.1
